requirements.txt
Django>=5.2.1
djangorestframework>=3.15.1
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.1
mysqlclient>=2.1.1
Pillow>=10.3.0
gunicorn>=20.1.0 